requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image==0.16.2
jupyter
# tensorflow-gpu==1.12.0
# python 3.6
keras==2.2.5
opencv-python
h5py
imgaug
# https://github.com/matterport/Mask_RCNN
# git+https://github.com/philferriere/cocoapi.git#egg=pycocotools^&subdirectory=PythonAPI